refactor: выделяет папку global для типов и констант
