+++ src/features/userMenu/ui/userMenu.tsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom';
 
-import { AppRoute } from '@/const';
+import { AppRoute } from '@/global/const';
 
 export function UserMenu() {
   return (
