+++ src/global/types/index.ts
@@ -1,0 +1,5 @@
+export type { UserType } from './user';
+export type { LocationType } from './location';
+export type { CityType } from './city';
+export type { FullOfferType, PreviewOfferType, OpenedOfferType } from './offer';
+export type { ReviewType } from './review';
