+++ src/global/types/location.ts
