+++ src/global/types/offer.ts
