+++ src/pages/404/ui/404.tsx
@@ -1,8 +1,9 @@
 import { Link } from 'react-router-dom';
 import { Helmet } from 'react-helmet-async';
 
+import { AppRoute } from '@/global/const';
+
 import styles from '../styles/styles.module.css';
-import { AppRoute } from '@/const';
 
 export function Page404() {
   return (
