--- src/shared/types/index.ts
@@ -1,5 +1,0 @@
-export type { UserType } from './components/user';
-export type { LocationType } from './components/location';
-export type { CityType } from './components/city';
-export type { FullOfferType, PreviewOfferType, OpenedOfferType } from './components/offer';
-export type { ReviewType } from './components/review';
